Added automatic databinding capabilities to the Chart Engine.
Three new methods have been added to the Chart Engine Generator class:

1- void Generator.bindData(IDataRowExpressionEvaluator expressionEvaluator, Chart chart, RunTimeContext rtc) throws ChartException
Binds data to the chart model using a row expression evaluator. The evaluator provides the ability to evaluate the expressions set in the chart on a row context.

2- void Generator.bindData(ResultSet resultSet, Chart chart, RunTimeContext rtc) throws ChartException
Binds a sql Resuset to a chart model. This is based on the assumption the column names of the resultset match exactly the data query definitions and other expressions set inside the chart model.

3- List Generator.getRowExpressions(Chart cm) throws ChartException
This retrieves all the row expressions stored in the chart model. This is useful to prepare a specific query for the chart.