- Summary: Added a checkbox to disable Live Preview
- Bugzilla Bug (s) Resolved:
121980

- Description:
121980 part fix: Errors in chart wizard when using scripted data source

With the new live preview feature of the chart builder, it could happen that some datasource/dataset problems cause some errors or unexpected behavior (like in this bug). We are always using cached data and revert automatically to sample data (non-live) in case of error, but it might not cover all possible cases, so it's better to give control to the user to enable/disable the live preview inside the chart builder. The fix would be to add a checkbox in the wizard to enable/disable live preview manually, by default set to enabled.

- Tests Description:

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Added:

- Files Edited: