- Summary: [Chart usability spec#2.4] UI refinement of Legend
- Bugzilla Bug (s) Resolved:
N/A

- Description:
Quotes chart UI usability spec #2.4:
As a common block, Legend is supposed to hide/display easily. It¡¯s better to display the checkbox for Legend visibility in the Legend subtask.
Format specifier of base series is applied to category values of Legend, so it¡¯s better to display it beside Legend text area.
Please note that this attribute is only available when users have set ¡°color by¡± as ¡°categories¡± in Series subtask. It will be disabled if the format specifier is invalid.

- Tests Description: Manual test
- Notes to Build Team: None
- Notes to Developers: None
- Notes to QA: None
- Notes to Documentation: None

- Files Edited:

- Files Added:

- Files Deleted