- Summary: Added a maximum check for the step number in fixed step case
 - Bugzilla Bug (s) Resolved:
150476, 168805

 - Description:
150476: BIRT 2.0 - OutOfMemoryException (Java Heap Space) :
Fixed by adding the maximum check for the step number in fixed step case. Now if setting a minor step in UI or API won't cause too many steps in scale computation, but throws an exception to indicate improper step or unit is set.
168805: Too many ticks notification when setting Filter condition to a scatter chart having been set scale [0001]:
Removed the check for step number in auto scale loop. Now if too many steps are produced in auto scale loop computation, skip the check.

In bubble size computation, replace an exception thrown by logging it.
- Tests Description:  Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA:  N/A

- Notes to Documentation:  N/A

- Files Edited:

- Files Added:

- Files Deleted