- Summary: Refactor and enhance chart to divide the execution of non-UI transaction and UI transaction of live preview. (#19011)
- TED Bug (s) Resolved: 19011[Partial]

- Description:
Since the cube initialization and cube data retrieving cost lots of time, now refactor and enhanced chart to divide the execution of non-UI transaction and UI transaction of live preview,there is a live preview thread to be created when chart builder shows up, this live preview thread is used to control the execution of divided live preview transactions. The cube initialization transaction (non-UI) will be executed in non-UI thread, and the chart render transaction (UI) will be executed in UI thread. It can avoid the UI freeze before first live preview when opening chart builder.

- Tests Description: Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA: N/A

- Notes to Documentation: N/A