- Summary: Chart ellipsis don't work when the chart axis are flipped(#31130)
- Bugzilla Bug (s) Resolved: 314708

- Description:
The cause of this bug is the axis label is too long and full fill the chart plot, so there is no space for chart series rendering. Now we use the same behavior with Excel to limit the length of axis label can't exceed half of chart width and add ellipsis in label tail.

- Tests Description: Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA: N/A

- Notes to Documentation: N/A